+++ src/platformAccessory.ts
@@ -6,7 +6,10 @@ import {Options, PythonShell} from 'python-shell';
 
 /* eslint-disable */
 const { exec } = require('child_process');
+const { path } = require('path');
 /* eslint-enable */
+
+const directory = process.cwd();
 
 export class WyzeThermostatAccessory {
   private service: Service;
@@ -19,7 +22,7 @@ export class WyzeThermostatAccessory {
   private dataTimeout = this.platform.config.newDataTimeout;
   private refreshIntervalID;
 
-  private p2stubs = this.platform.config.path2py_stubs;
+  private p2stubs = path.join(directory, 'py_helpers');
   private username = this.platform.config.username;
 
   private stateOff = this.platform.Characteristic.TargetHeatingCoolingState.OFF;
@@ -407,6 +410,10 @@ export class WyzeThermostatAccessory {
     //   // this.wyzeLog('Device STDERR: ' + stderr);
     // });
 
+    pyshell.on('error', (err) => {
+      this.wyzeLog('Python error getting thermostat states!');
+      this.wyzeLog(err);
+    });
 
     // end the input stream and allow the process to exit
     pyshell.end((err, code, signal) => {